Refactored tests in order to make future testing easier
